requirements.txt
python-dotenv>=1.0.0
langchain-openai>=0.3.0
requests>=2.28.0
pyyaml>=6.0